Use int32 instead of int64 as advised by VK docs
Use composition over inheritance with VkDate
Remove extra deserializers

+++ vk-api-core/src/main/kotlin/name/anton3/vkapi/tokens/Helpers.kt
@@ -8,14 +8,14 @@ val Token<*>.clientSecret: String?
         else -> null
     }
 
-val Token<UserGroupMethod>.peerId: Long
+val Token<UserGroupMethod>.peerId: Int
     get() = when (this) {
         is UserToken -> id
         is GroupToken -> -id
         else -> error("Impossible")
     }
 
-val Token<*>.peerId: Long?
+val Token<*>.peerId: Int?
     get() = when (this) {
         is UserToken -> id
         is GroupToken -> -id
