+++ vk-api-methods/src/main/kotlin/name/anton3/vkapi/methods/longpoll/events/IncomingMessageRead.kt
@@ -2,6 +2,6 @@ package name.anton3.vkapi.methods.longpoll.events
 
 data class IncomingMessageRead(
     override val eventType: Int,
-    val peerId: Long,
-    val localId: Long
+    val peerId: Int,
+    val localId: Int
 ) : LongPollEvent(eventType)
