+++ vk-api-methods/src/main/kotlin/name/anton3/vkapi/methods/longpoll/events/UserTypingInChat.kt
@@ -2,6 +2,6 @@ package name.anton3.vkapi.methods.longpoll.events
 
 data class UserTypingInChat(
     override val eventType: Int,
-    val userId: Long,
-    val chatId: Long
+    val userId: Int,
+    val chatId: Int
 ) : LongPollEvent(eventType)
